Ejercicio 12.1

+++ Ejercicio 12/Ejercicio 12/Program.cs
@@ -1,5 +1,5 @@
-﻿int numero = -1;
-int suma = 0;
+﻿double numero = -1;
+double suma = 0;
 int ordinales = 1;
 Console.WriteLine("Ingrese los números que quiera sumar e ingrese 0 para mostrar la suma final");
 Console.WriteLine();
@@ -7,7 +7,7 @@ do
 {
     Console.WriteLine($"Ingrese el {ordinales}° número");
     Console.WriteLine();
-    numero=int.Parse( Console.ReadLine() );
+    numero=double.Parse( Console.ReadLine() );
     Console.WriteLine();
     suma += numero;
     ordinales++;
